Segmented main component into small ones, folder made for components and helpers for fetching data
