+++ src/app/api/analyze/route.js
@@ -4,7 +4,6 @@ import { authOptions } from '../auth/[...nextauth]/route';
 
 export async function POST(req) {
 
-  /*
   let session;
 
     try {
@@ -14,9 +13,7 @@ export async function POST(req) {
         }
     } catch (error){
         return NextResponse.json({ message: 'Error en la sesión, debe iniciar sesión para utilizar este servicio'}, {status: 500});
-    }
-  */
- 
+    } 
     try {
       const formData = await req.formData();
       const file = formData.get('file');
