+++ src/app/service/components/EmotionsBar.jsx
